# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: starterpack
    description: Best API REST building with GoaDesign
    contact:
        name: Guillaume MORIN
        email: guillaume.morin@epitech.eu
        url: https://guillaume-morin.fr/
    license:
        name: GOA X GM
        url: https://guillaume-morin.fr/
    version: "1.0"
host: localhost:8088
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /v1/email-exist:
        post:
            tags:
                - auth
            summary: email-exist auth
            description: Check if email exist in database
            operationId: auth#email-exist
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: Email-ExistRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/AuthEmailExistRequestBody'
                    required:
                        - email
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/AuthEmailExistResponseBody'
                        required:
                            - exist
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/AuthEmailExistUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
...
    /v1/lost:
        post:
            tags:
                - auth
            summary: send-confirmation auth
            description: Check if email exist in database and send code by email to reset password
            operationId: auth#send-confirmation
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: Send-ConfirmationRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/AuthSendConfirmationRequestBody'
                    required:
                        - email
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/AuthSendConfirmationResponseBody'
                        required:
                            - exist
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/AuthSendConfirmationUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
    /v1/reset-password:
        put:
            tags:
                - auth
            summary: reset-password auth
            description: Reset forget password of the user with the correct confirm code
            operationId: auth#reset-password
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: Reset-PasswordRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/AuthResetPasswordRequestBody'
                    required:
                        - email
                        - code
                        - password
                        - confirm_password
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/AuthResetPasswordResponseBody'
                        required:
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/AuthResetPasswordUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
...
definitions:
    AuthEmailExistRequestBody:
        title: AuthEmailExistRequestBody
        type: object
        properties:
            email:
                type: string
                example: guillaume@gmail.com
                format: email
        example:
            email: guillaume@gmail.com
        required:
            - email
    AuthEmailExistResponseBody:
        title: AuthEmailExistResponseBody
        type: object
        properties:
            exist:
                type: boolean
                example: true
            success:
                type: boolean
                example: true
        example:
            exist: true
            success: true
        required:
            - exist
            - success
    AuthEmailExistUnknownErrorResponseBody:
        title: AuthEmailExistUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: false
        required:
            - err
            - success
            - error_code
    AuthResetPasswordRequestBody:
        title: AuthResetPasswordRequestBody
        type: object
        properties:
            code:
                type: string
                example: ZGI5E
                pattern: ^[A-Za-z0-9]+$
                minLength: 5
                maxLength: 5
            confirm_password:
                type: string
                description: Minimum 9 charactères / Chiffre Obligatoire
                example: JeSuisUnTest974
                pattern: \d
                minLength: 9
            email:
                type: string
                example: guillaume@gmail.com
                format: email
            password:
                type: string
                description: Minimum 9 charactères / Chiffre Obligatoire
                example: JeSuisUnTest974
                pattern: \d
                minLength: 9
        example:
            code: ZGI5E
            confirm_password: JeSuisUnTest974
            email: guillaume@gmail.com
            password: JeSuisUnTest974
        required:
            - email
            - code
            - password
            - confirm_password
    AuthResetPasswordResponseBody:
        title: AuthResetPasswordResponseBody
        type: object
        properties:
            success:
                type: boolean
                example: false
        example:
            success: true
        required:
            - success
    AuthResetPasswordUnknownErrorResponseBody:
        title: AuthResetPasswordUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: false
        required:
            - err
            - success
            - error_code
    AuthSendConfirmationRequestBody:
        title: AuthSendConfirmationRequestBody
        type: object
        properties:
            email:
                type: string
                example: guillaume@gmail.com
                format: email
        example:
            email: guillaume@gmail.com
        required:
            - email
    AuthSendConfirmationResponseBody:
        title: AuthSendConfirmationResponseBody
        type: object
        properties:
            exist:
                type: boolean
                example: false
            success:
                type: boolean
                example: true
        example:
            exist: false
            success: false
        required:
            - exist
            - success
    AuthSendConfirmationUnknownErrorResponseBody:
        title: AuthSendConfirmationUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: true
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: true
        required:
            - err
            - success
            - error_code
...
securityDefinitions:
    OAuth2_header_Authorization:
        type: oauth2
        description: Use OAuth2 to authenticate
        flow: application
        tokenUrl: /authorization
        scopes:
            api:read: Provides read access
    jwt_header_jwtToken:
        type: apiKey
        description: |-
            Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

            **Security Scopes**:
              * `api:read`: Read-only access
              * `api:write`: Read and write access
        name: jwtToken
        in: header